ibkr_quant/requirements.txt
ib_insync
pandas
numpy
python-dotenv
